pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "easyspectra"
version = "0.1.0"
description = "Hyperspectral and multispectral analysis toolkit with a graphical interface."
readme = "README.md"
requires-python = ">=3.10"

authors = [
    { name = "Matheus de Freitas Souza" }
]

dependencies = [
    "numpy",
    "matplotlib",
    "scikit-image",
    "opencv-python",
    "rasterio",
    "torch",
    "Pillow",
    "tifffile"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
easyspectra = "easyspectra.interface:main"

